feat(core): render --out-format for daemon transfers
Extends the remote --out-format rendering to the daemon transport, the
counterpart of the SSH push/pull work. Previously a daemon push wrote
the server's default preformatted line straight to stdout (a plain
FnMut(&str) sink), and a daemon pull passed no itemize callback at all,
so a custom --out-format template was ignored on both directions.

Both daemon drivers now route their client-visible rows through a shared
ItemizeEventSink (extracted from the SSH driver to remote/itemize_sink.rs
for reuse): with a custom --out-format the sink captures each row as a
metadata-bearing ClientEvent for the CLI's out-format renderer; otherwise
it writes the default line to stdout unchanged. The daemon push attaches
the collected events to its summary, and the daemon pull passes the sink
as the itemize callback (the receiver drains its buffered events into it,
wired in the transfer crate's receiver branch).

out_format_active is set once in apply_common_daemon_config, so both the
daemon generator (push) and receiver (pull) honour it; wants_client_output
fires under a custom --out-format even without -v/-i.

Verified against rsync 3.4.4 over a daemon connection: %o %n, %n, -i, and
[%i] %n (%l) render byte-for-byte identically for both push and pull.
