+++ crates/core/src/client/remote/mod.rs
@@ -34,6 +34,8 @@ pub(crate) mod flags;
 pub(crate) mod implied_source;
 /// Remote rsync `--server` invocation argument builder.
 pub mod invocation;
+/// Shared client-visible itemize sink for the remote transports.
+pub(crate) mod itemize_sink;
 /// `--info=`/`--debug=` server-argument construction (make_output_option).
 pub(crate) mod output_option;
 /// Remote-to-remote transfer via local proxy relay.
